Converter a requisição em JSON para InfluxDB 2.0 line protocol e gravar no banco de dados Fixes #9
