+++ tsdb/requirements.txt
@@ -1,3 +1,4 @@
 Flask==1.1.2
 influxdb-client==1.7.0
+python-dotenv==0.13.0
 requests==2.24.0
